🛠️ index.ts -> Updated createContextTheme function 🛠️ themeContext.ts -> Updated createContextTheme function 🛠️ index.tsx -> Updated createContextTheme function 🛠️ MessageHistory.tsx -> Updated import path for ConfigurationContext 🛠️ MessageInput.tsx -> Updated import path for ConfigurationContext 🟢 configurationContext.ts -> Renamed from utilities/configurationContext.ts 🟢 index.ts -> New file added for context 🛠️ index.tsx -> Updated import and usage of createContextConfiguration 🟢 IContextConfiguration.ts -> Renamed from IConfiguration.ts 🔴 utilities/index.ts
