+++ webview/message/src/components/MessageInput/MessageInput.tsx
@@ -7,7 +7,7 @@ import {
 import { Send16Regular, Mic24Regular, Mic24Filled } from '@fluentui/react-icons'
 import { FC, useCallback, useContext, useEffect, useRef, useState } from 'react'
 import { IMessageInputProps } from '../../interfaces'
-import { ConfigurationContext } from '../../utilities'
+import { ConfigurationContext } from '../../context'
 
 const useStyles = makeStyles({
   outerWrapper: {
