+++ webview/message/src/context/index.ts
@@ -1,0 +1,4 @@
+export {
+  ConfigurationContext,
+  createContextConfiguration,
+} from './configurationContext'
