+++ webview/message/src/interfaces/index.ts
@@ -3,4 +3,4 @@ export type { ICodeDocument } from './ICodeDocument'
 export type { IMessageHistoryProps } from './IMessageHistoryProps'
 export type { IMessageUtilityProps } from './IMessageUtilityProps'
 export type { IMessageInputProps } from './IMessageInputProps'
-export type { IConfiguration } from './IConfiguration'
+export type { IContextConfiguration } from './IContextConfiguration'
