+++ webview/message/src/utilities/index.ts
@@ -1,3 +1,2 @@
 export { vscode } from '../../../common/vscode'
 export { TextToSpeech } from './textToSpeech'
-export { ConfigurationContext } from './configurationContext'
